restructered for deployment
