--- back-end/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "back-end",
-  "version": "1.0.0",
-  "description": "back-end for note-mate",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
-  },
-  "author": "Clara Andrew-Wani",
-  "license": "ISC",
-  "dependencies": {
-    "bcrypt": "^1.0.3",
-    "cors": "^2.8.4",
-    "express": "^4.16.3",
-    "jsonwebtoken": "^8.2.0",
-    "mongoose": "^5.0.10"
-  }
-}
