+++ client/src/less/notesPage.less
