+++ package.json
@@ -1,19 +1,24 @@
 {
-  "name": "back-end-project-week",
+  "name": "back-end",
   "version": "1.0.0",
-  "description": "This week you will be building a backend for a note taking app called \"Lambda Notes.\"",
-  "main": "index.js",
+  "description": "back-end for note-mate",
+  "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/clarakosi/back-end-project-week.git"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node server.js",
+    "heroku-postbuild": "cd client && npm install --only=dev && npm install && npm run build"
   },
   "author": "Clara Andrew-Wani",
   "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/clarakosi/back-end-project-week/issues"
+  "dependencies": {
+    "bcrypt": "^1.0.3",
+    "cors": "^2.8.4",
+    "dotenv": "^5.0.1",
+    "express": "^4.16.3",
+    "jsonwebtoken": "^8.2.0",
+    "mongoose": "^5.0.10"
   },
-  "homepage": "https://github.com/clarakosi/back-end-project-week#readme"
+  "engines": {
+    "node": "9.2.1"
+  }
 }
